+++ README.md
@@ -1,4 +1,4 @@
-# LittleShellScripts
+# LittleRubyScripts
 
 **This is still a work in progress**
 
@@ -7,7 +7,7 @@ TODO: Delete this and the text above, and describe your gem
 ## Installation
 Install it yourself as:
 
-    $ gem install little_shell_scripts
+    $ gem install little_ruby_scripts
 
 ## Usage
 
@@ -21,7 +21,7 @@ To install this gem onto your local machine, run `bundle exec rake install`. To 
 
 ## Contributing
 
-1. Fork it ( https://github.com/[my-github-username]/little_shell_scripts/fork )
+1. Fork it ( https://github.com/[my-github-username]/little_ruby_scripts/fork )
 2. Create your feature branch (`git checkout -b my-new-feature`)
 3. Commit your changes (`git commit -am 'Add some feature'`)
 4. Push to the branch (`git push origin my-new-feature`)
