+++ lib/little_ruby_scripts.rb
@@ -1,0 +1,5 @@
+require "little_ruby_scripts/version"
+
+module LittleRubyScripts
+  # Your code goes here...
+end
