+++ lib/little_ruby_scripts/history.rb
