+++ lib/little_ruby_scripts/version.rb
@@ -1,0 +1,3 @@
+module LittleRubyScripts
+  VERSION = "0.1.0"
+end
