--- lib/little_shell_scripts.rb
@@ -1,5 +1,0 @@
-require "little_shell_scripts/version"
-
-module LittleShellScripts
-  # Your code goes here...
-end
