--- lib/little_shell_scripts/version.rb
@@ -1,3 +1,0 @@
-module LittleShellScripts
-  VERSION = "0.1.0"
-end
