+++ spec/little_ruby_scripts_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+describe LittleRubyScripts do
+  it 'has a version number' do
+    expect(LittleRubyScripts::VERSION).not_to be nil
+  end
+
+end
