--- spec/little_shell_scripts_spec.rb
@@ -1,8 +1,0 @@
-require 'spec_helper'
-
-describe LittleShellScripts do
-  it 'has a version number' do
-    expect(LittleShellScripts::VERSION).not_to be nil
-  end
-
-end
