+++ spec/spec_helper.rb
@@ -1,3 +1,3 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'little_shell_scripts/history'
-require 'little_shell_scripts'
+require 'little_ruby_scripts/history'
+require 'little_ruby_scripts'
